Split test pynini into multiple files
