+++ rustfst-tests-pynini/src/algorithms/inverse.rs
@@ -1,0 +1,26 @@
+use failure::Fallible;
+
+use rustfst::algorithms::invert;
+use rustfst::fst_traits::MutableFst;
+use rustfst::fst_traits::TextParser;
+use rustfst::semirings::Semiring;
+use rustfst::semirings::WeaklyDivisibleSemiring;
+
+use crate::TestData;
+
+pub fn test_invert<F>(test_data: &TestData<F>) -> Fallible<()>
+where
+    F: TextParser + MutableFst,
+    F::W: Semiring<Type = f32> + WeaklyDivisibleSemiring,
+{
+    // Invert
+    let mut fst_invert = test_data.raw.clone();
+    invert(&mut fst_invert);
+    assert_eq!(
+        test_data.invert,
+        fst_invert,
+        "{}",
+        error_message_fst!(test_data.invert, fst_invert, "Invert")
+    );
+    Ok(())
+}
