+++ rustfst-tests-pynini/src/algorithms/mod.rs
@@ -1,0 +1,12 @@
+pub mod arc_map;
+pub mod arcsort;
+pub mod connect;
+pub mod determinize;
+pub mod encode;
+pub mod inverse;
+pub mod project;
+pub mod properties;
+pub mod reverse;
+pub mod rm_epsilon;
+pub mod state_map;
+pub mod weight_pushing;
