+++ rustfst-tests-pynini/src/algorithms/rm_epsilon.rs
@@ -1,0 +1,26 @@
+use failure::Fallible;
+
+use rustfst::algorithms::{isomorphic, rm_epsilon};
+use rustfst::fst_impls::VectorFst;
+use rustfst::fst_traits::MutableFst;
+use rustfst::fst_traits::TextParser;
+use rustfst::semirings::Semiring;
+use rustfst::semirings::StarSemiring;
+use rustfst::semirings::WeaklyDivisibleSemiring;
+
+use crate::TestData;
+
+pub fn test_rmepsilon<F>(test_data: &TestData<F>) -> Fallible<()>
+where
+    F: TextParser + MutableFst,
+    F::W: 'static + Semiring<Type = f32> + WeaklyDivisibleSemiring + StarSemiring,
+{
+    // Remove epsilon
+    let fst_rmepsilon: VectorFst<_> = rm_epsilon(&test_data.raw).unwrap();
+    assert!(
+        isomorphic(&fst_rmepsilon, &test_data.rmepsilon)?,
+        "{}",
+        error_message_fst!(test_data.rmepsilon, fst_rmepsilon, "RmEpsilon")
+    );
+    Ok(())
+}
