+++ rustfst-tests-pynini/src/macros.rs
@@ -1,0 +1,8 @@
+macro_rules! error_message_fst {
+    ($fst_ref:expr, $fst:expr, $operation_name:expr) => {
+        format!(
+            "\nTest {} with pynini failing : \nREF = \n{}\nPRED = \n{}\n",
+            $operation_name, $fst_ref, $fst
+        )
+    };
+}
